+++ src/main/java/com/odeal/otomat/repository/ProductRepository.java
@@ -6,5 +6,5 @@ import org.springframework.stereotype.Repository;
 
 @Repository
 public interface ProductRepository extends JpaRepository<Product, Long> {
-    Product findBySlotNumber(Long slotNumber);
+    Product findBySlotNumber(int slotNumber);
 }
