+++ src/main/java/com/odeal/otomat/service/IProductService.java
@@ -14,6 +14,6 @@ public interface IProductService {
 
     Boolean delete(Long issue);
 
-    ProductDTO findBySlotNumber(Long slotNumber);
+    ProductDTO findBySlotNumber(int slotNumber);
 
 }
